todo_backend/todo_backend/urls.py
"""todo_backend URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/3.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path, include
from rest_framework.documentation import include_docs_urls
from rest_framework_swagger.views import get_swagger_view
from rest_framework import routers
from todo.urls import router as todo_router
from user_management.urls import router as user_router

router = routers.DefaultRouter()
router.registry.extend(todo_router.registry)
router.registry.extend(user_router.registry)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('', include(router.urls)),
    path('api-auth/', include('dj_rest_auth.urls')),
    path('api-user/', include('dj_rest_auth.urls')),
    path('api/docs/', include_docs_urls(title='Todo App API')),
    path('api/swagger-docs/', get_swagger_view(title='Todo App API'))
]
