requirements.txt
django==3.2.11
psycopg2==2.9.3
djangorestframework==3.13.1
importlib-metadata==4.10.1
markdown==3.3.6
zipp==3.7.0
django-filter==21.1
dj-rest-auth==2.2.2
django-cors-headers==3.11.0